backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-cli>=0.0.14",
    "fastapi[standard]>=0.119.1",
    "gitpython>=3.1.45",
    "jinja2>=3.1.6",
    "langchain>=1.0.2",
    "langchain-community>=0.4",
    "langfuse>=3.8.1",
    "litellm>=1.78.7",
    "markdown>=3.9",
    "prometheus-client>=0.23.1",
    "pydantic>=2.12.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-mock>=3.15.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.14.1",
    "tree-sitter>=0.25.2",
    "uvicorn>=0.38.0",
]

[tool.fastapi]
app = "src.api:app"